requirements.txt
ultralytics
python-dotenv
requests
tabulate
datumaro
pyyaml
cvat-sdk
Pillow
typer
